requirements.txt
python-dotenv
billboard.py
boxoffice_api
google-generativeai